move cpu, cuda and other device modules into kernels
